Ask Optiq: split the tool executor by what each tool touches
rocprofvis_ai_tools.cpp had grown to 2747 lines doing three jobs at once: the
dispatcher, the tools that drive the UI, and the tools that read the trace.
The two halves carry different hazards - a UI tool goes through OptiqActions
and answers in the same call, while a data tool has to reason about request
ids and result slots it shares with the rest of the app - so they are now
separate files that can be read and reviewed on their own.

- rocprofvis_ai_ui_tools.cpp: goto, show_panel, switch_tab, flow_arrows,
  annotate, bookmark, measure, reset_view, offer_next_steps. This is now the
  only file in the executor that names OptiqActions, so the read half cannot
  mutate the UI even by accident - a property that was previously only
  convention.
- rocprofvis_ai_data_tools.cpp: the eleven query tools, every formatter they
  use, the trace briefing and activity-chart helpers, and
  FinishAssistantFetch.
- rocprofvis_ai_tools.cpp keeps StartAssistantTool and the five helpers both
  halves need. Two of those are worth naming, because they are why goto and
  annotate are not purely UI tools: SelectedOrFullTimeRange, since annotate
  stores the window the user was looking at, and FindComputeKernel, since goto
  selects a kernel by the same loose name kernel_metrics accepts.
- rocprofvis_ai_tools_internal.h is private wiring. Each body file owns its
  own handler table and hands it over through an accessor, so the handlers
  keep internal linkage and adding a tool never touches the dispatcher: a
  schema entry, a body, and a line in that file's own table.

Bodies were moved verbatim - no behaviour change. Also drops four includes the
executor no longer used (spdlog, event manager, events, iomanip), and brings
the .agents/UI.md inventory back in step: it still described the module as
"four files" and had never been updated for the earlier tool_query and
tool_schema split. The UI.md hunks documenting ApplyPanelVisibilitySettings
and the shared-request-id rule belong with the previous commit and rode along
with this one.

Not build-verified: ROCPROFVIS_ENABLE_AGENTIC_PROFILING is off by default and
this machine has no mbedtls submodule checked out. Verified instead by
checking that all twenty tool names still resolve across the two tables, that
every moved symbol is defined exactly once, and that neither body file
references the other's helpers.
